Increase performance and improve mempool (#372)
Increase performance of the mempool by:
- Having the `mempool_state` and `blockchain` locks less time taken.
- Spawning the transaction verification task.

Fix possible issue with the `mempool_state` locking:

Before this change the read lock of the `mempool_state` was dropped
before taking a write lock to it to push a transaction. This may
lead to issues were a transaction could change the `mempool_state`
in between.
This change re-adds the upgradable RW lock of the `mempool_state`
which is now returned by the `verify_tx` function such that the
caller can upgrade the lock and push a transaction.