network: Integrate the message behaviour functionality into the connection pool
The message behaviour is currently in charge of creating peer objects, creating
and handling message dispatchers, and also closing connections. This is because
it needs to poll the peers sockets until completion when a connection closes.

This commit removes the message behaviour by integrating it's functionality
into the connection pool behaviour so that we can handle peer objects and
connections through a single behaviour. Also, it preps the way for removing
the message functionality all together and replacing it with libp2p's own
Request Response behaviour.