Increase the initial DB map size
Increase the initial DB map size from 1G to 1T.
Also increase the percentage threshold that defines when to resize
the DB to 100% such that the DB never gets resized.
This is to address issues resizing LMDB which is not well supported
and the recommendation is to use a large initial mapping size.
This will affect 32-bit systems and systems without sparse file
system support that are also written in the system requirements
section in the `README.md` file.

This closes #419.