Prevent panic in the block producer
Sometimes a block producer will keep running in parallel to a rebranch
operation which would lead to a panic as soon as this block producer
tries to query the blockchain (which is now in on the new branch
using its state (which is from the old branch).

This commit prevents the panic from happening and logs this situation as
a warning (because we should still check when this happens, since it could
also be due to a different reason than the one above and it should be
investigated too).