Add a blockchain trusted push function
Add a blockchain trusted push function such that when knowing the
source of a block the push can omit certain checks such as:
- VRF seed
- Block signature
- Transaction signature

This is particular useful for the validator which already verifies
the block so it can be directly pushed to the blockchain without
going to all this checks all over again.
This functionality is for now under a validator compilation feature
called `trusted_push`.