Add verification of staking transactions
Since the mempool performance optimizations, the mempool no longer
fully verifies a staking transaction.
While this needs a more performant solution, for the time being
this is adding staking transaction verifications after the
transactions are obtained from the mempool.
In case any transaction is dropped, the logic retries to fetch
more blocks from the mempool until an upper limit is reached or
the mempool returns no new transactions.

This fixes #415.