Use the correct prev view number when pushing a block
Currently, we're using the wrong previous view number when pushing the first
block after a macro block (because it should be the macro block's own view
number and we're using the next view number which is always 0), this
causes problems when reverting blocks (because there we're using the
right view number).

This fixes #423.