Fix incoming dial failure handling for connection pool behavior
Fix incoming dial failure handling for connection pool behavior.
Since the recent update of libp2p, [this
change](https://github.com/libp2p/rust-libp2p/pull/2191) introduces
state to dial request that can be transmitted in
`inject_dial_failure`. So there are a couple of cases where the
function is called but we can ignore the error since the connection
is still to be completed or is already connected.

This fixes #447.