Reduce the number of concurrent dial addresses to 1
Reduce the number of concurrent dial addresses to 1 in the network,
which is also the default hence the code won't set any value to it.
This was detected as the cause of several dials being launch when
issuing a peer dial and provoking error messages since only one of
the dial actions succeed while the other fails with messages
reporting incoming connection errors.

This fixes #446.