Fix devnet scripts to properly configure a spammer
Fix devnet scripts to properly configure a spammer.
Since the spammer has a hardcoded private key, when using a spammer
the genesis must have the account that the private key belongs to.
In order for that to happen, an optional argument was added to the
python `devnet_create` script to only generate the spammer
configuration files if the spammer option is specified.