Fix panic in consensus sync clustering
Fix panic trying to access the `epochs_ids` array out of bounds.
The access is done in `cluster_epoch_ids` and happens when the
received epoch IDs are from a peer that is behind our current
epoch ID. To fix this, the code will check for this case and
ignore the received epoch IDs since they come from a peer that
is behind our accepted state.

This fixes #490.