History Sync: Make sync queue to handle only weak references
Change `sync_queue` interface to avoid the need of returning strong
references and make the consumer handle when to upgrade an agent
reference.
This allows strong references of consensus agents to be held less
time and only get upgraded when it is absolutely required.
Also makes consistent the interface of `sync_queue` handling and
returning only weak reference counters.