Properly clean consensus agents from HistorySync
Add several changes to properly clean consensus agents from
`HistorySync`.
This includes these changes to `sync_queue`:
- Keep track of the peer ID of each consensus agent to properly
  identify it.
- Change `has_peer` function to properly detect when a peer has
  been added to the `SyncQueue`. Before, the detection was broken
  since it was using pointer comparison which only works for
  same instance detection.
- Add a function to remove peers from the `SyncQueue` such that
  when required a proper peer cleaning can be made.

This also includes an extra function to `HistorySync` to remove
peers from the different set of clusters present. This function is
also called upon the detection of the `PeerLeft` event such that
no stale peers are found in clusters while a peer was already been
removed from the list of `HistorySync` agents as it was happening
in #498.

This fixes #498.