Use different Type IDs for LevelUpdate messages
Use different Type IDs for `LevelUpdate` messages to allow the
network to properly decode an incoming message using the proper
type.
Since `LevelUpdateMessage` is a `struct` with generic elements,
it can't have a fixed Type ID since this is the mechanism the
network has to differentiate between types and then later do
a proper deserialization. As seen in #433, this may lead to
circumstances where the network fails to deserialize these type
of messages with the BLS signature deserialization being the most
notorious error.
This change changes the `LevelUpdateMessage` to not have a fixed
Type ID and use instead the one provided by the element implementing
the `AggregatableContribution` trait. This way the different
`LevelUpdateMessage` would have their own Type ID and then the
network would be able to properly decode them.

This fixes #433.