Properly implement gossipsub peer scoring
Properly implement gossipsub peer scoring by setting topic score
parameters after subscribing to a topic.
This allows the gossipsub scoring mechanism to properly account
the weights and penalties per topic as described in the gossipsub
v1.1 extension.
Also, avoid disconnecting a peer when it reaches a negative score
since this has these 2 downsides:
1. Interferes with the already present scoring mechanism in
   gossipsub.
2. Closing a connection to a peer doesn't prevent the peer to
   connect again as soon as the peer can.

Before this change it was seen that all peer scores were zero or a
negative value and that provoked:
1. A non functional peer scoring affecting internal gossipsub
   algorithms.
2. Peer disconnections since the code was closing connections to
   peers with negative scores.

This fixes #562.