Change the blake2b/blake2s hash functions to blake3
See https://blake3.io/. This should improve performance a little, as
blake3 is substantially faster than blake2b/blake2s. In a tiny
microbenchmark for a payload size of 32 bytes, it was 30% faster than
blake2b and 40% faster than blake2s.

TPS in the devnet test script varied so much across runs that no
definite improvement could be determined in the global context.