Don't rebuild nimiq-genesis for each devnet script run
Achieve this by reading the genesis block at runtime if
`NIMIQ_OVERRIDE_DEVNET_CONFIG` is set.

This unfortunately introduces a runtime dependency of `genesis` to
`nimiq-build-tools`. This could be fixed by either moving
`GenesisBuilder` out of `nimiq-build-tools` or making this
`NIMIQ_OVERRIDE_DEVNET_CONFIG` feature optional by introducing a cargo
feature for it.