Ignore duplicates in gossipsub publishing
Trying to tackle some of the warnings that happen when we start the
devnet. Duplicate blocks that are caught locally are no problem, so stop
warning about them.

Warnings that are suppressed (only the validator ones, the others come
from libp2p):
```
2022-03-11T23:47:21.610009Z WARN  validator            | Failed to publish block #384: Network(GossipsubPublish(Duplicate))
2022-03-11T23:47:28.620783Z WARN  behaviour            | Not publishing a message that has already been published. Msg-id 3133373630303630393431323431363133353136    , swarm task{peer_id=PeerId("12D3KooWCXJmuYWbiYp5aiWihYUJbLRr3nE4LY167z7KZkzwckqX")}
2022-03-11T23:47:28.620930Z WARN  validator            | Failed to publish block #512: Network(GossipsubPublish(Duplicate))
2022-03-11T23:47:30.884547Z WARN  behaviour            | Not publishing a message that has already been published. Msg-id 37323936363438363936303132313730333438    , swarm task{peer_id=PeerId("12D3KooWCXJmuYWbiYp5aiWihYUJbLRr3nE4LY167z7KZkzwckqX")}
```

I now just squelched all of the duplicate message warnings, would it be
preferrable to make that dependent on the topic and still warn about
proposals or transactions that are sent twice to the network?