Exit on errors in the devnet script
This e.g. stops the script if the Rust compiler returns an error
compiling the project.

Generally, having
```sh
set -o errexit
set -o nounset
set -o pipefail
```
on top of bash scripts is a good idea.

`errexit` exits the script with a non-zero exit status if any called
command returns a non-zero exit status (and is itself not inside an `if`
or similar). A few commands return a non-zero exit status on non-error
conditions, e.g. `grep`, which then might make stuff like `|| true` or
`|| test $? = 1` necessary to ignore all or specific non-zero exit
statuses.

`nounset` exits the script if an unset variable is used.

`pipefail` extends the `errexit` check to commands that are not the
final part of a pipeline.