Remove `ip` feature
This would have been a bit annoying to port, would have needed to
copy-paste the functions from the standard library. Luckily, the code
using these functions was dead, so we could instead remove the functions
calling `is_global`.

Since this is the final unstable `#![feature]` that was used, the
project now compiles on stable Rust!

Fixes #689.