Remove `Peer`, `MockPeer` struct and `Peer` trait
They didn't contain a lot of functionality anymore, but created a lot of
noise at the call sites. Use `PeerId`s instead, which are copyable and
can be easily stored.

The only remaining functionality of `Peer` was the `close` function
which got replaced by the `disconnect_peer` function on the `Network`
trait.