libp2p: Allow two connections per peer
Increase the maximum connections per peer from 1 to 2 and change
`NetworkBehavior` implementations to only report a newly joined peer
for the first connection only. An exception to this is the connection
pool behavior that needs to report to the `ConnectionHandler` every
connection and if it's inbound or outbound to properly build the
peer with the corresponding socket.
This avoids the issue reported in #741 hitting the connection limit
when two peers dial each other at the same time.

This fixes #741.