Improve the network default response
Improve the network default response by adding a field in the request
that identifies when the network should auto reply to such requests.
This is helpful specially in the aggregation cases where the
application may not be able to respond to the requests because:
- The receiver may be already dropped.
- The receiver may not be actively listening to the incoming
  requests.
In these cases the network does a best effort to deliver messages
(transmitted in form of requests) but the receiver may not be
actively getting them. The network auto replies to these messages
such that `Timeout` errors are not seen in the request/response
functionality.