history_sync: Use chunks for pushing history
Change the `push_history_sync` functionality in the `blockchain` to
receive `HistoryTreeChunks` progressively instead of a full epoch.
This includes the following changes:
- Remove `push_history_sync` method.
- Add `start_history_sync` method that only takes a `MacroBlock` and
  returns a `HistoryPusher` instance.
- Add `HistoryPusher` struct with associated functions to:
  - Add history chunks in the form of `HistoryTreeChunk`s.
  - Commit changes when no more history chunks are expected.
Also changed the consensus `cluster` and `sync_stream` functionality
to use the new APIs and call the appropriate functions as soon as new
history chunks are received.

This addresses #896.