Add a new type of blocks: skip blocks.
Add skip block as a new type of block. This also required changes in
several places in the code mainly in `blockchain` and `validator`
crates such that this type of block can be produced and accepted by
the blockchain.

Also remove the concept of view change from the code and rename the
view change delay in favor of the block producer timeout. This
removes the `view_number` field in either macro and micro blocks.
For micro blocks it also implies that there are no view change
proofs as justification.

This fixes #867.