Use `block_number` as offset for micro block slots
Use `block_number` as offset for getting the slot owner for micro
blocks.
Using a constant as before doesn't work for skip blocks since the
VRF seed is carried over. This causes that the slot owner of a block
after a skip block keeps being the same and then only skip blocks
are produced for the rest of that epoch.