Introduce `round` for macro blocks
Introduce a new field in macro blocks (`round`) to register the
round number a proposal was made. This replaces the old concept of
`view_number` where the round was stored and then used to verify the
VRF seed once the slot owner was obtained.
This was required since the `round` noted in the `TendermintProof`
is the block assembler `round` which differes from the one in the
proposal and is required for checking the signature in the proof.