history_sync: Use macro block terminated chunks
Use macro block terminated chunks to push history.
For this, the accumulated size of extended transactions is used to
keep an account of it and signal which checkpoint macro blocks must
be kept.
History sync will request such intermediate checkpoints (if they
exist) along with the corresponding history instead of requesting
a full epoch to then push it to the blockchain.

This resolves #896.