history_sync: Move check of the history store root
Move the check of the history store root when pushing history using
`blockchain`'s `push_history_sync` to the `extend_history_sync`
function. This is to check the history store root when adding the
transactions to the history store but before committing the DB
transaction such that it can be aborted in case of failure.
Also this avoids the usage of memory for checking the history store
since it becomes noticeable when using an already populated history
store and large epochs.