Introduced provided and needed services
When peers are connecting to each other, now they have two different services flags:
Provided and Needed.
The first one is used to signal to other peers which services are supported, while the other one
is used to identify if another is interesting to me:
Another peer is interesting to me if my needed services are contained in his provided services.

Signed-off-by: viquezclaudio <viquezclaudio@gmail.com>