Add a block separation time
Add a block separation time and change the micro block producer
within the `validator` to check when a block needs to be either:
- Immediately produced because the expected block time for a block
has already passed.
- Delayed in order to comply with the block separation time.

The block separation time will be taken as a point for the reward
function later to incentivize validators to produce blocks within
this time.