consensus: Change `BlockLiveSync` to always request macro bodies
Change `BlockLiveSync` to always request macro bodies. For this the
`include_body` flag is changed to `include_micro_bodies`.
This affects mainly `RequestMissingBlocks` and its handler since now
the handler must detect the type of block being requested and include
the body depending on the `include_micro_bodies` flag if the block is
a micro block. If the block is a macro block then bodies are always
returned.