consensus: Change requests to always request macro bodies (#1177)
* consensus: Change `BlockLiveSync` to always request macro bodies

Change `BlockLiveSync` to always request macro bodies. For this the
`include_body` flag is changed to `include_micro_bodies`.
This affects mainly `RequestMissingBlocks` and its handler since now
the handler must detect the type of block being requested and include
the body depending on the `include_micro_bodies` flag if the block is
a micro block. If the block is a macro block then bodies are always
returned.

* consensus: Change `BlockRequest` to always send macro bodies

Change the `BlockRequest` to always send macro bodies. For that, the
`include_micro_bodies` flag was added to the request and the handler
only sends micro bodies when set but macro bodies are always sent.