Create a new `database-value` crate
Create a new `database-value` crate and move the `IntoDatabaseValue`,
`FromDatabaseValue` and `AsDatabaseBytes` traits to this new crate.
This is to avoid `database` dependencies in the `primitives` crates
or in any other crate that only requires these traits.
This is required because the `database` trait depends in `mdbx` which
doesn't compile for wasm.

This fixes #987.