network-libp2p: Fix close connections for non-IP addresses
Fix issue in the connection pool behaviour when closing a connection
for a peer that doesn't have the `ip` protocol in its `Multiaddr`.
The issue was that the code wasn't marking the connection as closed
in the behaviour connection state and it wasn't matching the state
on the rest of the network.