Simplify the computation of the txn count
The accumulated transaction count that is now included as part of ChainInfo is needed for generation txns proofs.
However, for the current epoch, all that is needed is the accumulated txn count since the beggining of the current epoch.
This is because the history tree operates on a per epoch basis (and not on a per batch basis)
So there is no need to compute, save and expose the accumulated txn count of each checkpoint block.