Get rid of forked `ed25519-zebra` dependency
It was forked to make the `s` field of `SigningKey` public, it's the
computed scalar for the ed25519 signature. Since its computation is
specified in RFC 8032, we can compute it ourselves and get rid of the
fork.

The actual computation is cheap enough (0.3µs) so that the double
computation isn't a problem.