Use the history tree length for generating a proof
Use the history tree length for generating a transaction proof since
this is what the history tree interface expects when generating a
proof for a previous state of the tree.
Also rename the `cum_ext_tx_count` field of `ChainInfo` to
`history_tree_len` and change the blockchain code to properly set it
based on the current length of the history tree.