network-libp2p: Allow only a connection per peer
Allow only a connection per peer by making the connection pool
behaviour to close a connection if we see more than a connection to
the same peer.
This also fixes instabilities in the
`connection_stress_and_reconnect` test that were produced because the
test was hanging waiting to receive events for a closed connection
but they were sometimes never emitted because there were duplicated
connections to a peer and only one of them was being closed.

This closes #832.