network-libp2p: Deterministically close duplicated connections
Change the connection pool behaviour to deterministically choose
what connections must be closed when there are multiple to the same
peer. The behaviour will close the connection if there are any
previous connection and if the peer ID value is less or equal than
the local (own) peer ID.

This fixes #1412.