network-libp2p: Fix panic in the discovery handler
Fix panic in the discovery handler that was produced when the handler
outbound and inbound connections were resolved before processing the
event for setting the peer address that gets reported on a connection
established event in the behaviour.
The panic signature was an `expect` for the peer address that results
in this message: `Address should have been resolved`.