consensus: Add a way of getting staking contract elements
Add a way of getting staking contract elements to a remote endpoint.
For this `RemoteDataStore` was implemented in consensus such that it
builds the appropriate `TrieProofRequest` to then verify the received
proofs and then deserialize the data appropriately.
The implementation differs for now on `wasm` and non `wasm` targets
since in `wasm` right now everything runs in a single thread and we
can't block it. Therefore, in `wasm` the implementation is for now
leaking the `StakingContractStore` instead of just using sync
interface of `RemoteDataStore` for the staking contract.