Request transaction by hash
Currently we have the possibility to request a transaction proof by hash and block number.
In this PR a new variant is introduced, that adds the possibility to request the proof by providing only the block number.
In other words, the proving block is determined from the server side and not the client side.