network-libp2p: Finish DHT queries as soon as one result is available
Finish `DHT` queries as soon as one result is reported. This
replicates the previous behavior where for getting a record from the
`DHT` a quorum of one was used. This behavior got changed in #1240
and was provoking several queries received for the same record and
thus unexpected error messages such as the ones reported in #1499.
A TODO was added such that this policy of using a quorum of 1 is
revisited later and in case of changing this a proper implementation
of verifying and selecting the best value must be added.

This fixes #1499.