Use a plain client config passed from main thread to worker
Similar to the last commit, we need to serialize the config to pass it from the main thread to the worker, which mean introducing a plain config version into which the config gets encoded in the main thread and decoded from in the worker.

Furthermore, this change has the advantage that `Client.create()` now simply takes a plain object with all properties optional, making it possible to pass a config object with the right properties in directly, without _having_ to go through the ClientConfiguration builder.