Convert TransactionBuilder into a standalone static class
We cannot use TransactionBuilder client-connected, as the client and the TransactionBuilder (or at least the transactions that it outputs) live in different thread WASM memories, and accessing them from each other between these contexts is not (easily) possible.

Unfortunately, this means for now users have to specify the validity start height and network ID manually, for which I added accessor methods on the client.