handel: Allow empty levels and avoid invalid IDs in levels
Allow empty levels and avoid invalid IDs in levels.
This is useful to account for level complete appropriately instead
of waiting for aggregations from peers that will never arrive. Also
this avoid accounting internal state for peers that doesn't exist
such as the state in the `Level` code for selecting the next peers
for sending updates.