Implement Javascript files for example and package with updated client APIs
For the NPM package, we create two builds, one for bundlers (that can use import from package names), and one for plain web use, that needs to specify the exact file that is imported, both in the main thread and in the webworker.

This also updates the demo (now in example/) module.js file to use the updated client API, using the TransactionBuilder statically and passing serialized (to hex in this case) transactions when sending them. No other changes are required in the demo file.

The `Client` class/object defined in dist/client-proxy.js _shadows_ the `class Client` exported by the worker types and adheres to that one's static API. For the user it is transparent.