web-client: Fix client log level setting
Fix the setting of the log level for the client. The `primitives`
feature was missing the `camelCase` `serde` directive and the `Tsify`
derive directive for the `ClientConfiguration` struct.
Also made the `primitives` and `client` features the default features
and changed the script to disable the default features.
Also changes the script to use `set -e` to avoid the usage of `&&` in
each line.