transaction: Fix `AnyHash` deserialization
Fix `AnyHash` deserialization for human readable deserializers.
Human readable deserializers regularly use `visit_map` for
deserializing `struct`s so it is actually unnecessary to detect them
in the visitor and only implementing `visit_seq` would result in an
error deserializing. This is why this change implements `visit_map`
to the visitor.