blockchain: Fix verifying history setting for the genesis
Fix the verifying history setting for a corner case where a DB is
loaded and the block number is still the genesis one (no history yet).
This was provoking that `can_verify_history` to be set as false as the
genesis history root doesn't match the hash of the history store.