Modify transaction struct itself when calling sign() on it
Instead of returning a signed copy. I think this API is more intuitive.

I also don't want to do both, modify self and return a signed copy (because it's not possible to just return self, that is first-of-all a compiler error, and secondly would create a second instance in jS-land anyway), as that would transparently create a copy of the variable, which is unexpected behavior for users of the library and will lead only to errors.