Correctly check the signer of the VRF seed for macro block proposals.
Previously the vrf seed would be checked against the proposer for the round that is given in valid_round. That is however not necessarily correct as that round could already be a re-proposed proposal, making that proposer not the signer of the seed.

Now the round given in the header is taken for seed verification and proposals which do not correctly state a VR if they should are being rejected.