Improve DHT publishing
Previously, the own DHT entry was published on every new epoch. That means it might try to publish before the DHT finished setting up. Also, it republished the DHT although we configured the library to automatically do that every 60 seconds.
This PR adds an event to the Network that is fired when DHT is done bootstrapping. The validator listens for this event and publishes the own record upon receiving it.