network-libp2p: Re-implement request response codec
Re-implement the request-response protocol codec such that the
network can react to empty requests or responses that are originated
from reading EOF at the beginning of the socket.
This also gets rid of the prefixed length read and writes that were
introduced to the codec in the upgrade to libp2p 0.53.1 and instead
uses a fixed size prefix (`u64`).
This introduces a backwards incompatibility in the request-response
protocol.
This fixes #804.