network-libp2p: Add the `autonat` bahaviour
Add the `autonat` behaviour to detect when a peer is behind a private
`NAT`. If in the other hand a peer has a public reachable IP, then
the behaviour will fire the `ExternalAddrConfirmed` event and will
put the `kad` behaviour (DHT) into server mode.
This removes the previous setting where the DHT was being set always
in server mode regardless of the node reachability except for the
memory transport where Autonat doesn't work (only works with IPs).