network: Implement selection of DHT (kad) records
Change the `network-interface` to receive parameters that implements
the `PartialOrd` trait such that the network can compare them and be
able to select the best one.
Change the `network-libp2p` implementation to expect a quorum of 3
answers before accepting a DHT record and add logic to replicate the
best answer to outdated peers.
This closes #1531.