network-libp2p: Add some throttling for re-connections
Make the `ConnectionState` aware of the desired number of connections
such that when doing `housekeeping` it resets the connections marked
as down after 1s when the number of connections is below the desired
number. This changes how eagerly connections are re-tried when the
number of connected peers is less than the desired value.