Add direction to `RequestMissingBlocks`.
Direction::Forward lacks the property of being able to provide non main chain blocks. If looking up a block on an inferior chain and wanting to connect it to the main chain, Direction::Backward now provides that functionality.
In this commit it is unused but will be used in a future commit.

The body of `handle_backward` is a previously replaced version of `handle`. Todo was added as it was as it seems still relevant to include at some point.