Add deterministic builds
Introduces bit-for-bit reproducible aka deterministic builds. The
`build/README.md` has a usage guide and details on what this type of
security control protects from.
The main impact of this change is on the way the release is built. It
introduces a small build system which leverages make and docker to
build `core-rs-albatross` in a deterministic manner. The new build
process preserves the current release path `./target/release`. As a
result the `./Dockerfile` which is present in the repo will use the
deterministically built code by default, if the deterministic build
flow is used.