validator-network: Separate state lock in different locks
Change the `validator-network` implementation to separate the lock
around the state into different locks for the different subcomponents
of the network state such when modifying a part of the state doesn't
require to block reads to the other subcomponents of the state. Since
they are independent from each other, they should be allowed to be
modified independently as well.