network-libp2p: Fix mishandling of `FinishedWithNoAdditionalRecord`
Fix the mishandling of `FinishedWithNoAdditionalRecord` since we
could get such event even if the `query.finish` isn't called (which
the network is calling once `Quorum` records have been received).
This mishandling was causing some of the queries to hang indefinitely
since no error nor result was being sent to the application layer.
Now the network always pushes the result to the application layer
once the `FinishedWithNoAdditionalRecord` is received.