consensus: Use election macro head for validating batch sets
Use election macro head for validating batch sets instead of using
the macro head. This is because we could ask for a batch set of an
epoch containing checkpoint macro blocks we already know and in this
case, the verification of batch sets would fail since it may contain
checkpoint macro blocks we already know and that has a block number
lower than our macro head.