network-libp2p: Make the `ConnectionState` service aware
Make the `ConnectionState` aware of the `Service`s that our node
requires to be able to differentiate between regular connected peers
and those connected peers that are interested to the node (that
contains the node's required services).
This allows the network to always try to maintain
`desired_peer_count` connections that also meet our required services
instead of just counting any connection.
This fixes #2313.