utils: Fix some TOCTOU issues
https://en.wikipedia.org/wiki/Time-of-check_to_time-of-use

Generally, it's better to try and open a file and check the error to see
if it did not exist, rather than trying to figure out the existence of a
file using different means.

Do the same for creating parent directories. First, try creating a file.
Only if that fails due to the missing parent directory, use
`std::fs::create_dir_all` instead of `std::fs::create_dir` to create the
parent directory. This function does not fail if the directory already
exists. Then, try creating the file again.