Correct serialization of leaf value for merkle path computation
When computing a multisig merkle path for SignatureProofs, the leaf value was still the `PublicKey` enum, whose serialization contains a leading variant ID, whereas the merkle tree values where already serialized without the variant ID. The version without the variant ID is correct, to keep backwards-compatibility with pre-webauthn multisigs.

Turns out, when computing a merkle path where the leaf to proof is not contained in the values, an empty path is generated.