Dont move incompatible peers to live sync too early.
If peers are classified as incompatible and moved to live sync before the validity sync is complete,
this creates issues where missing blocks can be pushed from the state sync, even though,
we cannot verify the history root yet.
With this change, we first complete the validity sync and after that, we move the incompatible peers to live sync.