primitives: Change the `supply_at` function to not require genesis TS
Change the `supply_at` function in the policy module within the
primitives subcrate to depend only on the PoW genesis timestamp
defined as a constant such that the supply _follows_ the PoW supply
curve as a continuation. This also changes the signature of this
function since the PoS genesis timestamp is no longer required.