blockchain: Update chain info when reverting a block
Update chain info when reverting a block from the blockchain such
the block is moves from the `pushed` block table to the `stored`
block table. This is necessary since once reverted there is going
to be no history for such block and the block needs to be fully
stored. Once a batch finalizes, the block can be removed (as it
happens with all blocks in the `stored` block table).