pow-migration: Do not use `Policy` functions to calculate epoch numbers
Change the blockchain history store interface to add a function
(`add_to_history_for_epoch`) that receives the epoch number instead
of relying on `Policy` to calculate it. This is useful in the history
migration for `PoW` since the genesis block number wouldn't be known
until the migration is over.